Fix load-order dependency
This fixes an `unitialized constant CC::Service::GitHubPullRequests`,
which was happening if `CC::Service::GithubPullRequests::Presenter` was loaded
before `CC::Service::GithubPullRequests`. Loading order is non-deterministic and
can differ depending on OS, since services are loaded via `Dir[]`).

Going forward, services that need a presenter class can require the presenter
themselves.
